fix: changed create user endpoint to return UserProfileDto

+++ src/main/java/com/itesm/application/dto/CreateUserDto.java
@@ -21,7 +21,7 @@ public class CreateUserDto {
     private String password;
 
     @JsonProperty("role_id")
-    private Integer roleId;
+    private Byte roleId;
 
     @JsonProperty("suburb_id")
     private Integer suburbId;
